Fix delete button in `echo-http-counter` demo
Previously, interacting with the delete button in the demo would cause a
`SyntaxError: Unexpected token 'return'` error. This is because the
`before` attribute of the delete button had a value of `return
confirm('Are you sure?')`, resulting in a newly created function
resulting in `return return confirm('Are you sure?')`.

This commit removes the redundant "return" of the delete button's
`before` attribute, which resolves the previously mentioned error.